+++ CHANGES.md
@@ -1,3 +1,6 @@
+* Add missing `#ctcp_type` and `#dcc_type` accessors
+* Make `#type, `#ctcp_type`, and `#dcc_type` return symbols, not strings
+
 0.1.0
 -----
 * Initial release